+++ zng/builder/colbuilder.go
@@ -1,4 +1,4 @@
-package proc
+package builder
 
 import (
 	"errors"
